ref(api): Keep Entity mutations generic
Restore POST /entities with required kind and keep PATCH /entities/{entity} as the shared update boundary. Reserve the five kind-specific collections for browse operations.
