--- apps/server/src/orpc/contracts/blenders/create.ts
@@ -1,7 +1,0 @@
-import { entityKindCreateContract } from "../entityKinds/create";
-
-export default entityKindCreateContract({
-  operationId: "createBlender",
-  path: "/blenders",
-  summary: "Create a blender",
-});
