--- apps/server/src/orpc/contracts/bottlers/create.ts
@@ -1,7 +1,0 @@
-import { entityKindCreateContract } from "../entityKinds/create";
-
-export default entityKindCreateContract({
-  operationId: "createBottler",
-  path: "/bottlers",
-  summary: "Create a bottler",
-});
