--- apps/server/src/orpc/contracts/brands/create.ts
@@ -1,7 +1,0 @@
-import { entityKindCreateContract } from "../entityKinds/create";
-
-export default entityKindCreateContract({
-  operationId: "createBrand",
-  path: "/brands",
-  summary: "Create a brand",
-});
