--- apps/server/src/orpc/contracts/companies/create.ts
@@ -1,7 +1,0 @@
-import { entityKindCreateContract } from "../entityKinds/create";
-
-export default entityKindCreateContract({
-  operationId: "createCompany",
-  path: "/companies",
-  summary: "Create a company",
-});
