--- apps/server/src/orpc/contracts/distilleries/create.ts
@@ -1,7 +1,0 @@
-import { entityKindCreateContract } from "../entityKinds/create";
-
-export default entityKindCreateContract({
-  operationId: "createDistillery",
-  path: "/distilleries",
-  summary: "Create a distillery",
-});
