+++ apps/server/src/orpc/mock/contract.ts
@@ -3,30 +3,26 @@ import activityList from "@peated/server/orpc/contracts/activity/list";
 import login from "@peated/server/orpc/contracts/auth/login";
 import badgeDetails from "@peated/server/orpc/contracts/badges/details";
 import badgeUserList from "@peated/server/orpc/contracts/badges/user-list";
-import blenderCreate from "@peated/server/orpc/contracts/blenders/create";
 import blenderList from "@peated/server/orpc/contracts/blenders/list";
 import bottleGroupBottles from "@peated/server/orpc/contracts/bottleGroups/bottles";
 import bottleGroupDetails from "@peated/server/orpc/contracts/bottleGroups/details";
-import bottlerCreate from "@peated/server/orpc/contracts/bottlers/create";
 import bottlerList from "@peated/server/orpc/contracts/bottlers/list";
 import bottleDetails from "@peated/server/orpc/contracts/bottles/details";
 import bottleList from "@peated/server/orpc/contracts/bottles/list";
 import bottlePriceList from "@peated/server/orpc/contracts/bottles/prices/list";
 import bottleTags from "@peated/server/orpc/contracts/bottles/tags";
-import brandCreate from "@peated/server/orpc/contracts/brands/create";
 import brandList from "@peated/server/orpc/contracts/brands/list";
 import changeList from "@peated/server/orpc/contracts/changes/list";
 import collectionBottleDelete from "@peated/server/orpc/contracts/collections/bottles/delete";
 import collectionBottleList from "@peated/server/orpc/contracts/collections/bottles/list";
 import collectionBottleUpdate from "@peated/server/orpc/contracts/collections/bottles/update";
 import commentList from "@peated/server/orpc/contracts/comments/list";
-import companyCreate from "@peated/server/orpc/contracts/companies/create";
 import companyList from "@peated/server/orpc/contracts/companies/list";
 import countryDetails from "@peated/server/orpc/contracts/countries/details";
 import countryList from "@peated/server/orpc/contracts/countries/list";
-import distilleryCreate from "@peated/server/orpc/contracts/distilleries/create";
 import distilleryList from "@peated/server/orpc/contracts/distilleries/list";
 import entityCatalog from "@peated/server/orpc/contracts/entities/catalog";
+import entityCreate from "@peated/server/orpc/contracts/entities/create";
 import entityDetails from "@peated/server/orpc/contracts/entities/details";
 import entityList from "@peated/server/orpc/contracts/entities/list";
 import eventList from "@peated/server/orpc/contracts/events/list";
@@ -74,7 +70,7 @@ export const mockContract = {
     bottles: bottleGroupBottles,
     details: bottleGroupDetails,
   },
-  blenders: { create: blenderCreate, list: blenderList },
+  blenders: { list: blenderList },
   bottles: {
     details: bottleDetails,
     list: bottleList,
@@ -83,15 +79,15 @@ export const mockContract = {
     },
     tags: bottleTags,
   },
-  bottlers: { create: bottlerCreate, list: bottlerList },
-  brands: { create: brandCreate, list: brandList },
+  bottlers: { list: bottlerList },
+  brands: { list: brandList },
   changes: {
     list: changeList,
   },
   comments: {
     list: commentList,
   },
-  companies: { create: companyCreate, list: companyList },
+  companies: { list: companyList },
   collections: {
     bottles: {
       delete: collectionBottleDelete,
@@ -103,9 +99,10 @@ export const mockContract = {
     details: countryDetails,
     list: countryList,
   },
-  distilleries: { create: distilleryCreate, list: distilleryList },
+  distilleries: { list: distilleryList },
   entities: {
     catalog: entityCatalog,
+    create: entityCreate,
     details: entityDetails,
     list: entityList,
   },
