+++ openspec/changes/model-entity-kinds-and-ownership/proposal.md
@@ -9,8 +9,8 @@ as Diageo owning Lagavulin.
 
 - Keep one Entity record for each whisky identity or owning company.
 - Give each Entity one kind: Brand, Distillery, Bottler, Blender, or Company.
-- Expose each kind as its own top-level API collection. Keep a read-only
-  generic Entity collection for cross-kind selectors.
+- Expose each kind as its own top-level browse collection. Keep the generic
+  Entity API for cross-kind selection, creation, and shared Entity mutations.
 - Find Brand, Bottler, and Distiller use from Bottle links instead of copying
   those values onto Entity.
 - Let an Entity point to one current owner. Owner links can form a company
@@ -21,9 +21,9 @@ as Diageo owning Lagavulin.
   becomes required.
 - **BREAKING**: replace Entity `type` with `kind` in server, web, CLI,
   classifier, and generated API contracts. Use dedicated Brand, Distillery,
-  Bottler, Blender, and Company endpoints for browse and create operations.
-  Keep `GET /entities` for cross-kind selection and remove generic Entity
-  creation. Bottle fields do not store or filter by Entity roles.
+  Bottler, Blender, and Company endpoints for browse operations. Keep the
+  generic Entity API for selection, create, update, and other shared Entity
+  operations. Bottle fields do not store or filter by Entity roles.
 
 ## Capabilities
